Add files via upload
FINALLY got around to updating the Multinode chat.  It now displays like an IRC with chat type line at bottom and put comment above and scroll upwards per line of text.  Still not 100%, but at least it works.  If anyone wants to tweak it great.  Had a hell of a time getting it to not double space locally when remote worked, and vise versa.   But finally got it.